infra/src/scripts/internal/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "internal"
version = "0.0.1"
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Ahmad Khidir", email = "khidirahmad05@gmail.com" },
]

dependencies = [
  "aws-lambda-powertools",
  "pydantic",
  "boto3",
  "python-jose[cryptography]",
  "requests",
]

[tool.setuptools]
package-dir = {"" = "src"}
